+++ changelog.md
@@ -1,6 +1,7 @@
 ### Next release
 
 - [Steevan](https://github.com/steevanb) Add `--all-commands` to ignore removed commands configuration.
+- [Steevan](https://github.com/steevanb) Symfony 5 dependencies allowed.
 
 ### [1.1.1](../../compare/1.1.0...1.1.1) - 2016-05-29